EDU-724 Move media player related components into a dedicated folder
