+++ src/components/file-preview.js
@@ -4,13 +4,13 @@ import prettyBytes from 'pretty-bytes';
 import MiniPager from './mini-pager.js';
 import { message, Tooltip } from 'antd';
 import Logger from '../common/logger.js';
-import MediaPlayer from './media-player.js';
 import { useTranslation } from 'react-i18next';
 import LiteralUrlLink from './literal-url-link.js';
 import { useService } from './container-context.js';
 import { handleError } from '../ui/error-helper.js';
 import { useDateFormat } from './locale-context.js';
 import mimeTypeHelper from '../ui/mime-type-helper.js';
+import MediaPlayer from './media-player/media-player.js';
 import ClientConfig from '../bootstrap/client-config.js';
 import { getResourceType } from '../utils/resource-utils.js';
 import FileTextFilledIcon from './icons/files/file-text-filled-icon.js';
