+++ src/components/media-player/main-track-editor.yml
