+++ src/components/media-player/media-player-controls.less
