+++ src/components/media-player/media-player-progress-bar.less
