+++ src/components/media-player/media-range-selector.less
