+++ src/components/media-player/media-volume-slider.less
@@ -1,8 +1,8 @@
-.VolumeSlider {
+.MediaVolumeSlider {
   display: flex;
   align-items: center;
 }
 
-.VolumeSlider-slider {
+.MediaVolumeSlider-slider {
   width: 60px;
 }
