+++ src/components/media-player/timeline.less
