+++ src/components/media-player/track-mixer.yml
