+++ src/plugins/audio/audio-display.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import MediaPlayer from '../../components/media-player.js';
 import ClientConfig from '../../bootstrap/client-config.js';
 import { useService } from '../../components/container-context.js';
 import CopyrightNotice from '../../components/copyright-notice.js';
 import { sectionDisplayProps } from '../../ui/default-prop-types.js';
+import MediaPlayer from '../../components/media-player/media-player.js';
 import { MEDIA_SCREEN_MODE, MEDIA_SOURCE_TYPE } from '../../domain/constants.js';
 
 function AudioDisplay({ content }) {
