+++ src/plugins/interactive-media/interactive-media-editor.js
@@ -3,20 +3,20 @@ import classNames from 'classnames';
 import { useTranslation } from 'react-i18next';
 import validation from '../../ui/validation.js';
 import cloneDeep from '../../utils/clone-deep.js';
-import Timeline from '../../components/timeline.js';
 import { removeItemAt } from '../../utils/array-utils.js';
-import MediaPlayer from '../../components/media-player.js';
 import ClientConfig from '../../bootstrap/client-config.js';
 import React, { Fragment, useEffect, useState } from 'react';
 import DeleteButton from '../../components/delete-button.js';
 import MarkdownInput from '../../components/markdown-input.js';
 import InteractiveMediaInfo from './interactive-media-info.js';
+import Timeline from '../../components/media-player/timeline.js';
 import { useService } from '../../components/container-context.js';
 import { Button, Divider, Form, Input, Spin, Tooltip } from 'antd';
 import { sectionEditorProps } from '../../ui/default-prop-types.js';
-import MainTrackEditor from '../../components/main-track-editor.js';
 import { useNumberFormat } from '../../components/locale-context.js';
+import MediaPlayer from '../../components/media-player/media-player.js';
 import ObjectWidthSlider from '../../components/object-width-slider.js';
+import MainTrackEditor from '../../components/media-player/main-track-editor.js';
 import { MEDIA_SCREEN_MODE, MEDIA_SOURCE_TYPE } from '../../domain/constants.js';
 import { formatMediaPosition, getFullSourceUrl } from '../../utils/media-utils.js';
 import { CheckOutlined, LeftOutlined, PlusOutlined, RightOutlined } from '@ant-design/icons';
