+++ src/plugins/video/video-display.js
@@ -1,10 +1,10 @@
 import React from 'react';
-import MediaPlayer from '../../components/media-player.js';
 import ClientConfig from '../../bootstrap/client-config.js';
 import { MEDIA_SOURCE_TYPE } from '../../domain/constants.js';
 import { useService } from '../../components/container-context.js';
 import CopyrightNotice from '../../components/copyright-notice.js';
 import { sectionDisplayProps } from '../../ui/default-prop-types.js';
+import MediaPlayer from '../../components/media-player/media-player.js';
 
 function VideoDisplay({ content }) {
   const clientConfig = useService(ClientConfig);
