+++ src/styles/main.less
@@ -22,7 +22,6 @@
 @import '../components/alert.less';
 @import '../components/markdown.less';
 @import '../components/news-tab.less';
-@import '../components/timeline.less';
 @import '../components/room-card.less';
 @import '../components/rooms-tab.less';
 @import '../components/pages/user.less';
@@ -40,18 +39,15 @@
 @import '../components/pages/index.less';
 @import '../components/pages/login.less';
 @import '../components/pages/users.less';
-@import '../components/track-mixer.less';
 @import '../components/color-picker.less';
 @import '../components/pdf-document.less';
 @import '../components/file-preview.less';
 @import '../components/used-storage.less';
 @import '../components/files-viewer.less';
-@import '../components/media-player.less';
 @import '../components/tag-selector.less';
 @import '../components/abc-notation.less';
 @import '../components/pages/search.less';
 @import '../components/image-editor.less';
-@import '../components/volume-slider.less';
 @import '../components/document-card.less';
 @import '../components/grid-selector.less';
 @import '../components/control-panel.less';
@@ -83,7 +79,6 @@
 @import '../components/files-grid-viewer.less';
 @import '../components/files-list-viewer.less';
 @import '../components/edit-control-panel.less';
-@import '../components/media-player-track.less';
 @import '../components/ui-language-dialog.less';
 @import '../components/admin/settings-tab.less';
 @import '../components/document-info-cell.less';
@@ -91,26 +86,31 @@
 @import '../components/pages/create-import.less';
 @import '../components/default-page-header.less';
 @import '../components/pages/reset-password.less';
-@import '../components/media-range-selector.less';
+@import '../components/media-player/timeline.less';
 @import '../components/default-page-template.less';
 @import '../components/history-control-panel.less';
 @import '../components/document-import-table.less';
 @import '../components/files-upload-overview.less';
-@import '../components/media-player-controls.less';
 @import '../components/not-supported-section.less';
 @import '../components/dialog-footer-buttons.less';
 @import '../components/plugin-selector-dialog.less';
-@import '../components/multitrack-media-player.less';
 @import '../components/document-metadata-modal.less';
 @import '../components/admin/storage-plans-tab.less';
-@import '../components/media-player-progress-bar.less';
+@import '../components/media-player/track-mixer.less';
+@import '../components/media-player/media-player.less';
 @import '../components/never-scrolling-text-area.less';
 @import '../components/localization/language-icon.less';
 @import '../components/pages/complete-registration.less';
 @import '../components/pages/browser-not-supported.less';
 @import '../components/pages/complete-password-reset.less';
+@import '../components/media-player/media-player-track.less';
+@import '../components/media-player/media-volume-slider.less';
+@import '../components/media-player/media-range-selector.less';
 @import '../components/localization/country-flag-and-name.less';
 @import '../components/pages/room-membership-confirmation.less';
+@import '../components/media-player/media-player-controls.less';
+@import '../components/media-player/multitrack-media-player.less';
+@import '../components/media-player/media-player-progress-bar.less';
 
 
 // Pure CSS components
